Move CoreUtils authentication/permissions related functions into their own class
