+++ src/Classes/MyRadio/MyRadioMenu.php
@@ -9,16 +9,12 @@ namespace MyRadio\MyRadio;
 
 use \MyRadio\Config;
 use \MyRadio\MyRadioException;
+use \MyRadio\MyRadio\AuthUtils;
+use \MyRadio\MyRadio\CoreUtils;
 
 /**
  * Abstractor for the MyRadio Menu
- *
- * @author Lloyd Wallis <lpw@ury.org.uk>
- * @version 20130930
  * @package MyRadio_Core
- * @uses \CacheProvider
- * @uses \Database
- * @uses \CoreUtils
  */
 class MyRadioMenu
 {
@@ -105,7 +101,7 @@ class MyRadioMenu
 
         return $items;
     }
-    
+
     /**
      * Takes a $url database column entry, and breaks it into its components
      * @param  String $url A database-fetched menu item URL
@@ -129,7 +125,7 @@ class MyRadioMenu
     private function userHasPermission($item)
     {
         return empty($item['action']) or
-            CoreUtils::requirePermissionAuto($item['module'], $item['action'], false);
+            AuthUtils::requirePermissionAuto($item['module'], $item['action'], false);
     }
 
     /**
