+++ src/Classes/MyRadioTwig.php
@@ -5,6 +5,7 @@ namespace MyRadio;
 use \MyRadio\Config;
 use \MyRadio\Database;
 use \MyRadio\MyRadio\CoreUtils;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\MyRadioMenu;
 use \MyRadio\MyRadioException;
 use \MyRadio\MyRadioError;
@@ -168,7 +169,7 @@ class MyRadioTwig implements \MyRadio\Iface\TemplateEngine
      */
     public function render()
     {
-        if ((defined('AUTH_SHOWERRORS') && CoreUtils::hasPermission(AUTH_SHOWERRORS))
+        if ((defined('AUTH_SHOWERRORS') && AuthUtils::hasPermission(AUTH_SHOWERRORS))
              || Config::$display_errors) {
             $this->addVariable('phperrors', MyRadioError::$php_errorlist);
             if (isset($_SESSION)) { //Is the DB working?
