+++ src/Controllers/Mail/optin.php
@@ -8,6 +8,7 @@
  */
 
 use \MyRadio\MyRadioException;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_List;
 use \MyRadio\ServiceAPI\MyRadio_User;
@@ -17,7 +18,7 @@ if (!isset($_REQUEST['list'])) {
 }
 
 if (isset($_REQUEST['memberid'])) {
-    CoreUtils::requirePermission(AUTH_EDITANYPROFILE);
+    AuthUtils::requirePermission(AUTH_EDITANYPROFILE);
     $user = $_REQUEST['memberid'];
 } else {
     $user = -1;
