+++ src/Controllers/MyRadio/a-timeslotSignin.php
@@ -8,12 +8,13 @@
  * @package MyRadio_Core
  */
 
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_Timeslot;
 
 $ts = MyRadio_Timeslot::getInstance($_REQUEST['timeslotid']);
 if ($ts->getSeason()->getShow()->isCurrentUserAnOwner()
-        or CoreUtils::hasPermission(AUTH_EDITSHOWS)) {
+        or AuthUtils::hasPermission(AUTH_EDITSHOWS)) {
     $data = $ts->getSigninInfo();
     CoreUtils::dataToJSON($data);
 } else {
