+++ src/Controllers/NIPSWeb/manage_library.php
@@ -7,13 +7,14 @@
  * @package MyRadio_NIPSWeb
  */
 
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_User;
 use \MyRadio\NIPSWeb\NIPSWeb_ManagedPlaylist;
 use \MyRadio\NIPSWeb\NIPSWeb_ManagedUserPlaylist;
 
 $template = 'NIPSWeb/manage_library.twig';
-if (CoreUtils::hasPermission(AUTH_UPLOADMUSICMANUAL)) {
+if (AuthUtils::hasPermission(AUTH_UPLOADMUSICMANUAL)) {
     $template = 'NIPSWeb/manage_library_manual.twig';
 }
