+++ src/Controllers/Profile/edit.php
@@ -8,15 +8,16 @@
  */
 
 use \MyRadio\MyRadioException;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_User;
 use \MyRadio\ServiceAPI\MyRadio_Photo;
 
 // Set if trying to view another member's profile page
-if (isset($_REQUEST['profileedit-memberid']) && CoreUtils::hasPermission(AUTH_EDITANYPROFILE)) {
+if (isset($_REQUEST['profileedit-memberid']) && AuthUtils::hasPermission(AUTH_EDITANYPROFILE)) {
     $user = MyRadio_User::getInstance($_REQUEST['profileedit-memberid']);
 
-} elseif (isset($_REQUEST['memberid']) && CoreUtils::hasPermission(AUTH_EDITANYPROFILE)) {
+} elseif (isset($_REQUEST['memberid']) && AuthUtils::hasPermission(AUTH_EDITANYPROFILE)) {
     $user = MyRadio_User::getInstance($_REQUEST['memberid']);
 
 } else {
