+++ src/Controllers/Profile/view.php
@@ -10,6 +10,7 @@
  * @package MyRadio_Profile
  */
 
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_User;
 use \MyRadio\ServiceAPI\MyRadio_TrainingStatus;
@@ -30,7 +31,7 @@ if ($user->isOfficer()) {
     $userData['email'] = $user->getPublicEmail();
 }
 
-if (CoreUtils::hasPermission(AUTH_VIEWOTHERMEMBERS)) {
+if (AuthUtils::hasPermission(AUTH_VIEWOTHERMEMBERS)) {
     $userData['email'] = $user->getEmail();
     $userData['eduroam'] = $user->getEduroam();
     $userData['local_alias'] = $user->getLocalAlias();
@@ -50,41 +51,29 @@ if ($user->getID() === $visitor->getID() or $visitor->hasAuth(AUTH_EDITANYPROFIL
     $template->addVariable(
         'editurl',
         '<a href="'
-        .CoreUtils::makeURL(
-            'Profile',
-            'edit',
-            ['memberid' => $user->getID()]
-        )
+        .CoreUtils::makeURL('Profile', 'edit', ['memberid' => $user->getID()])
         .'">Edit Profile</a>'
     );
 }
-if (CoreUtils::hasPermission(AUTH_IMPERSONATE)
+if (AuthUtils::hasPermission(AUTH_IMPERSONATE)
     && ($user->hasAuth(AUTH_BLOCKIMPERSONATE) === false
-        or CoreUtils::hasPermission(AUTH_IMPERSONATE_BLOCKED_USERS))) {
+        or AuthUtils::hasPermission(AUTH_IMPERSONATE_BLOCKED_USERS))) {
     $template->addVariable(
         'impersonateurl',
         '<a href="'
-        .CoreUtils::makeURL(
-            'MyRadio',
-            'impersonate',
-            ['memberid' => $user->getID()]
-        )
+        .CoreUtils::makeURL('MyRadio', 'impersonate', ['memberid' => $user->getID()])
         .'">Impersonate User</a>'
     );
 }
-if (CoreUtils::hasPermission(AUTH_LOCK)) {
+if (AuthUtils::hasPermission(AUTH_LOCK)) {
     $template->addVariable(
         'lockurl',
         '<a href="'
-        .CoreUtils::makeURL(
-            'Profile',
-            'lock',
-            ['memberid' => $user->getID()]
-        )
+        .CoreUtils::makeURL('Profile', 'lock', ['memberid' => $user->getID()])
         .'">Disable Account</a>'
     );
 }
-if (CoreUtils::hasPermission(AUTH_MARKPAYMENT)) {
+if (AuthUtils::hasPermission(AUTH_MARKPAYMENT)) {
     $template->addVariable('can_mark_payments', true);
 }
 $template->render();
