+++ src/Controllers/Scheduler/showPhoto.php
@@ -8,6 +8,7 @@
  */
 
 use \MyRadio\MyRadioException;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_Show;
 
@@ -18,7 +19,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $show = MyRadio_Show::getInstance($data['show_id']);
     //Require this is the user's show or the user can edit any show
     if (!$show->isCurrentUserAnOwner()) {
-        CoreUtils::requirePermission(AUTH_EDITSHOWS);
+        AuthUtils::requirePermission(AUTH_EDITSHOWS);
     }
 
     $show->setShowPhoto($data['image_file']['tmp_name']);
@@ -35,7 +36,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
     //Require this is the user's show or the user can edit any show
     if (!$show->isCurrentUserAnOwner()) {
-        CoreUtils::requirePermission(AUTH_EDITSHOWS);
+        AuthUtils::requirePermission(AUTH_EDITSHOWS);
     }
 
     MyRadio_Show::getPhotoForm()
