+++ src/Models/MyRadio/actionPermissionsForm.php
@@ -1,5 +1,9 @@
 <?php
 
+use \MyRadio\MyRadio\AuthUtils;
+use \MyRadio\MyRadio\MyRadioForm;
+use \MyRadio\MyRadio\MyRadioFormField;
+
 /**
  *
  * @todo Proper Documentation
@@ -55,7 +59,7 @@ $form->addField(
                         'text' => 'GLOBAL ACCESS'
                     ]
                 ],
-                CoreUtils::getAllPermissions()
+                AuthUtils::getAllPermissions()
             )
         ]
     )
